requirements/base.txt
coverage==3.7.1
Django==1.7.1
django-appconf==1.0.2
django-author==0.1
django-autoslug==1.7.2
django-bootstrap3==2.5.5
django-choices==1.1.12
django-dynamic-fixture==1.6.5
django-extensions==1.3.2
django-filter==0.7
django-fsm==2.2.0
django-geoposition==0.2.3
django-grappelli==2.5.3
django-guardian==1.1.1
django-nose==1.2
django-override-settings==1.2
django-phonenumber-field==0.2a3
Django-Select2==4.2.2
django-selectable==0.9.0
django-tables2==0.15.0
django-user-accounts==1.3.1
importlib==1.0.4
mock==1.0.1
mocker==1.1.1
MySQL-python==1.2.5
nose==1.3.0
phonenumbers==7.7.2
Pillow==3.4.2
pytz==2016.7
setuptools-git==1.1
six==1.7.3
sorl-thumbnail==11.12.1b0
sqlparse==0.2.2
Werkzeug==0.9.4
